requirements.txt
djangorestframework
django-cors-headers
pillow
djoser
pytils
django-ckeditor
channels==2.4.0
channels-redis==3.2.0
daphne==2.5.0
yandex_checkout
django-cleanup